Update createCabin Api - to create or edit a cabin

+++ the-wild-oasis/src/services/apiCabins.js
@@ -14,44 +14,55 @@ export async function getCabins() {
     return data;
 }
 
-export async function createCabin(newCabin) {
-
-    //https://qewyniuqkyeaifamyeaz.supabase.co/storage/v1/object/public/cabin-images/cabin-002.jpg
+export async function createEditCabin(newCabin, id) {
+    const hasImagePath = Boolean(newCabin?.image?.startsWith?.(supabaseUrl));
 
     const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll('/', ''); //create a unique name for each image and remove '/' to avoid problems with supabase
 
-    const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+    const imagePath = hasImagePath ? newCabin.image : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-    // 1. Create a cabin
-    const { data, error } = await supabase
-        .from('cabins')
-        .insert([{ ...newCabin, image: imagePath }])
-        .select();
+    // 1. Create/Edit a cabin
+    let query = supabase.from('cabins');
+
+    // A) Create a cabin (if there is NOT id)
+    if (!id) {
+        query = query.insert([{ ...newCabin, image: imagePath }]);
+    }
+
+    // B) Edit (if there is id)
+    if (id) {
+        query = query.update({ ...newCabin, image: imagePath }).eq('id', id)
+    }
+
+    const { data, error } = await query.select().single();
+
 
     if (error) {
         console.error(error);
         throw new Error('Cabin can not be created!');
     }
 
-    console.log(data)
+    // Prevent uploading plain text.
+    if (!hasImagePath) {
 
-    // 2. Upload the image in cabin-images bucket
-    const { error: storageError } = await supabase
-        .storage
-        .from('cabin-images')
-        .upload(imageName, newCabin.image);
+        // 2. Upload the image in cabin-images bucket
+        const { error: storageError } = await supabase
+            .storage
+            .from('cabin-images')
+            .upload(imageName, newCabin.image);
 
-    // 3.Delete the cabin if there is a storageError(the image is uploaded successfully);
-    if (storageError) {
-        await supabase
-            .from('cabins')
-            .delete()
-            .eq('id', data.id);
 
-        console.error(storageError);
-        throw new Error('Cabin image could not be uploaded and the cabin was not created!');
+        // 3.Delete the cabin if there is a storageError(the image is NOT uploaded successfully);
+        if (storageError) {
+            await supabase
+                .from('cabins')
+                .delete()
+                .eq('id', data.id);
+
+            console.error(storageError);
+            throw new Error('Cabin image could not be uploaded and the cabin was not created!');
+        }
     }
-
     return data;
 
 }
